Refactor adaptive sections sample
